removed compilation error from service utils and made score type to INTEGER

+++ src/main/java/com/acko/insuredassetcredibility/models/KeyFactorDataScore.java
@@ -16,6 +16,6 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class KeyFactorDataScore {
     KeyFactorsData keyFactorsData;
-    Double score;
+    Integer score;
 
 }
